+++ src/test/java/com/tagmycode/sdk/DbServiceTest.java
@@ -3,6 +3,8 @@ package com.tagmycode.sdk;
 import com.tagmycode.sdk.model.Language;
 import com.tagmycode.sdk.model.Property;
 import com.tagmycode.sdk.model.Snippet;
+import org.h2.jdbc.JdbcSQLException;
+import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
@@ -12,7 +14,7 @@ import support.MemDbService;
 import java.io.IOException;
 import java.sql.SQLException;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.*;
 import static org.mockito.Mockito.*;
 
 
@@ -21,8 +23,31 @@ public class DbServiceTest extends BaseTest {
 
     @Before
     public void initSpyDbService() throws SQLException {
-        dbServiceSpy = spy(new MemDbService());
-        dbServiceSpy.initialize();
+        dbServiceSpy = spy(new MemDbService()).initialize();
+    }
+
+    @After
+    public void shutDownDb() throws IOException {
+        dbServiceSpy.close();
+    }
+
+    @Test
+    public void newInstanceOfDbHasCorrectSchemaVersion() throws SQLException {
+        MemDbService memDbService = new MemDbService("another_db");
+        assertFalse(memDbService.isCurrentSchemaVersion());
+
+        memDbService.initialize();
+
+        assertTrue(memDbService.isCurrentSchemaVersion());
+    }
+
+    @Test
+    public void setSchemaVersion() throws SQLException {
+        dbServiceSpy.propertyDao().deleteBuilder().delete();
+
+        dbServiceSpy.setCurrentSchemaVersion();
+
+        assertTrue(dbServiceSpy.isCurrentSchemaVersion());
     }
 
     @Test
@@ -82,6 +107,24 @@ public class DbServiceTest extends BaseTest {
     }
 
     @Test
+    public void dropAllTables() throws Exception {
+        Language language = resourceGenerate.aLanguage();
+        dbServiceSpy.languageDao().createOrUpdate(language);
+        assertLanguagesCount(dbServiceSpy, 1L);
+
+        dbServiceSpy.dropAllTables();
+
+        try {
+            dbServiceSpy.languageDao().countOf();
+            fail("Expected exception");
+        } catch (JdbcSQLException ignore) {
+        }
+        for (Class aClass : dbServiceSpy.getTableClasses()) {
+            verify(dbServiceSpy, times(2)).dropTable(aClass);
+        }
+    }
+
+    @Test
     public void createAndReadAllModelTypes() throws Exception {
         for (Class aClass : dbServiceSpy.getTableClasses()) {
             verify(dbServiceSpy, times(1)).createTableIfNotExists(aClass);
@@ -100,6 +143,7 @@ public class DbServiceTest extends BaseTest {
 
     @Test
     public void testPropertiesTable() throws SQLException {
+        dbServiceSpy.clearAllTables();
         dbServiceSpy.propertyDao().createOrUpdate(new Property("key", "value"));
         Property readProperty = dbServiceSpy.propertyDao().queryForId("key");
         assertEquals("value", readProperty.getValue());